Software/requirements.txt
# Core dependencies for laptop analysis server
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.11.0
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
pillow>=10.0.0

# Raspberry Pi specific (install only on Pi)
# picamera2>=0.3.12

